refactor: Controller에서 HttpResponse에 HttpStatus를 주입하도록 수정
