+++ src/main/java/configure/PathConstants.java
@@ -5,4 +5,6 @@ public class PathConstants {
     public static final String STATIC_URI_PREFIX = "./static";
 
     public static final String ERROR_PAGE_PATH = "/error.html";
+
+    public static final String DEFAULT_REDIRECT_LOCATION = "/index.html";
 }
