+++ src/main/java/controller/AbstractController.java
@@ -1,7 +1,6 @@
 package controller;
 
 import webserver.request.HttpRequest;
-import webserver.request.requestline.QueryParams;
 import webserver.response.HttpResponse;
 
 import java.io.IOException;
@@ -10,13 +9,10 @@ import java.net.URISyntaxException;
 public abstract class AbstractController implements Controller {
 
     void doGet(final HttpRequest request, final HttpResponse response) throws IOException, URISyntaxException {
-        // TODO : do something with query string params
-        QueryParams queryParams = request.findQueryParams();
-
         response.makeResponse(request);
     }
 
     void doPost(final HttpRequest request, final HttpResponse response) throws IOException, URISyntaxException {
-
+        response.makeResponse(request);
     }
 }
